+++ src/components/views/StealthField/stealthBoard.js
@@ -1,0 +1,82 @@
+import React, { Component } from "react";
+import TransitionGroup from "react-transition-group/TransitionGroup";
+import StealthBars from "./stealthBars";
+import { Row } from "reactstrap";
+
+const limit = 0.05;
+const factor = 0.005;
+
+export default class StealthBoard extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      systems: props.systems || null
+    };
+    this.looping = true;
+    this.loop = this.loop.bind(this);
+    window.requestAnimationFrame(this.loop);
+  }
+  componentDidMount() {
+    this.looping = true;
+  }
+  componentWillUnmount() {
+    this.looping = false;
+  }
+  componentWillReceiveProps(nextProps) {
+    if (nextProps.systems && !this.state.systems) {
+      // We only need to initialize the state
+      this.setState({
+        systems: nextProps.systems.filter(
+          s => typeof s.stealthFactor === "number"
+        )
+      });
+    }
+  }
+  loop(currentTime) {
+    if (this.looping) {
+      window.requestAnimationFrame(this.loop);
+    } else {
+      return;
+    }
+    if (Math.round(currentTime) % 2 !== 0) return;
+    if (!this.props.systems) return;
+    const systemsState = this.state.systems;
+    const systemsProps = this.props.systems;
+    if (!systemsState || !systemsProps) return;
+    this.setState({
+      systems: systemsState
+        .filter(s => typeof s.stealthFactor === "number")
+        .map(s => {
+          const propSys = systemsProps.find(ps => ps.id === s.id);
+          let sign = Math.sign(Math.random() - 0.5);
+          if (Math.abs(s.stealthFactor - propSys.stealthFactor) > limit) {
+            sign = -1 * Math.sign(s.stealthFactor - propSys.stealthFactor);
+          }
+          let stealthFactor = Math.min(
+            1,
+            Math.max(0, s.stealthFactor + sign * Math.random() * factor)
+          );
+
+          return {
+            id: s.id,
+            name: s.name,
+            type: s.type,
+            stealthFactor
+          };
+        })
+    });
+  }
+  render() {
+    const { stealthField } = this.props;
+    const { systems } = this.state;
+    return (
+      <Row className="stealth-board">
+        <TransitionGroup>
+          {[StealthBars]
+            .filter(s => stealthField.state)
+            .map(Comp => <Comp key={Comp.name} systems={systems} />)}
+        </TransitionGroup>
+      </Row>
+    );
+  }
+}
